refactor: move nav to sections and `components/common` to `components`
