--- src/components/common/index.ts
@@ -1,4 +1,0 @@
-export { Header } from "components/common/Header";
-export { ParseHTML } from "components/common/ParseHTML";
-export { To } from "components/common/To";
-export { Tooltip } from "components/common/Tooltip";
