+++ src/sections/projects/project/details/Forks.tsx
@@ -1,6 +1,5 @@
-import Detail from "sections/projects/project/details/Detail";
-
 import { ForksIcon } from "commons/icons";
+import Detail from "sections/projects/project/details/Detail";
 
 export const Forks = ({ forks }: { forks: number }): JSX.Element | null => (
     <Detail detail={forks}>
