+++ src/sections/projects/project/details/Stars.tsx
@@ -1,6 +1,5 @@
-import Detail from "sections/projects/project/details/Detail";
-
 import { StargazersIcon } from "commons/icons";
+import Detail from "sections/projects/project/details/Detail";
 
 export const Stars = ({ stargazers }: { stargazers: number }): JSX.Element | null => (
     <Detail detail={stargazers}>
