+++ src/sections/projects/project/index.tsx
@@ -2,7 +2,7 @@ import { fadeChild } from "commons/animations";
 import { GITHUB } from "commons/config";
 import { ProjectProps } from "commons/graphql";
 import theme from "commons/theme";
-import { Header, ParseHTML, To } from "components/common";
+import { Header, ParseHTML, To } from "components";
 import { Forks, Stars } from "sections/projects/project/details";
 
 import { motion } from "framer-motion";
